pyproject.toml
[build-system]
requires      = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "table15"
version = "0.0.2"
description = "Table 1.5 is a Python application that can generate a table that is adjunct to a typical Table 1 (association statistics). Table 1.5 goes beyond static association by analyzing the impact that a change in each single feature has to changes in the outcome."
readme = "README.md"
authors = [{ name="Rajas Kale", email="rajas.kale@carelon.com" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["Table1.5"]
requires-python = ">=3.7"
dependencies = [
    "pandas >= 0.25",
    "scikit-learn >= 1.1.2",
    "keras >= 2.10.0",
    "scikeras >= 0.9.0",
    "numpy >= 1.23.3",
    "shap >= 0.41.0",
    "PyYAML >= 6.0"
]

[project.urls]
Homepage = "https://https://github.com/KaleRP/table15"

[project.scripts]
table15 = "table15.__main__:main"